% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties.R
\name{addresses}
\alias{addresses}
\alias{from}
\alias{to}
\alias{cc}
\alias{bcc}
\alias{reply}
\alias{sender}
\title{Add address fields to message}
\usage{
from(msg, addr = NULL)

to(msg, ...)

cc(msg, ...)

bcc(msg, ...)

reply(msg, addr = NULL)

sender(msg, addr = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{addr}{Single address.}

\item{...}{Addresses.}
}
\value{
A message object.
}
\description{
Add address fields to message
}
\examples{
msg <- envelope()

# Populating the From field.
msg \%>\% from("craig@gmail.com")

# Populating the To field.
msg <- envelope()
msg \%>\% to("bob@gmail.com, alice@yahoo.com")
msg \%>\% to("bob@gmail.com", "alice@yahoo.com")
msg \%>\% to(c("bob@gmail.com", "alice@yahoo.com"))

# Populating the Cc field.
msg <- envelope()
msg \%>\% cc("bob@gmail.com, alice@yahoo.com")
msg \%>\% cc("bob@gmail.com", "alice@yahoo.com")
msg \%>\% cc(c("bob@gmail.com", "alice@yahoo.com"))

# Populating the Bcc field.
msg <- envelope()
msg \%>\% bcc("bob@gmail.com, alice@yahoo.com")
msg \%>\% bcc("bob@gmail.com", "alice@yahoo.com")
msg \%>\% bcc(c("bob@gmail.com", "alice@yahoo.com"))

# Populating the Reply-To field.
msg <- envelope()
msg \%>\% reply("gerry@gmail.com")

# Populating the Sender field.
msg <- envelope()
msg \%>\% sender("on_behalf_of@gmail.com")
}
