% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{append.envelope}
\alias{append.envelope}
\title{Append children to message}
\usage{
\method{append}{envelope}(x, child)
}
\arguments{
\item{x}{Message object}

\item{child}{A child to be appended}
}
\description{
Append children to message
}
