% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{as.address}
\alias{as.address}
\title{Create an address object}
\usage{
as.address(address)
}
\arguments{
\item{address}{An email address.}
}
\value{
An \code{address} object.
}
\description{
Create an address object
}
\examples{
as.address("gerry@gmail.com")
}
