% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{as.character.envelope}
\alias{as.character.envelope}
\title{Create formatted message.}
\usage{
\method{as.character}{envelope}(x, ..., details = TRUE)
}
\arguments{
\item{x}{A message object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A formatted message object.
}
\description{
Accepts a message object and formats it as a MIME document.
}
