% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{as.character.MIME}
\alias{as.character.MIME}
\title{Convert MIME object to character vector}
\usage{
\method{as.character}{MIME}(x, ...)
}
\arguments{
\item{x}{MIME object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Convert MIME object to character vector
}
