% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{as.character.vctrs_address}
\alias{as.character.vctrs_address}
\title{Convert address object to character}
\usage{
\method{as.character}{vctrs_address}(x, ...)
}
\arguments{
\item{x}{A vector of \code{address} objects.}

\item{...}{}
}
\value{
A character vector.
}
\description{
Convert address object to character
}
