% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add attachments to a message object}
\usage{
attachment(msg, path, cid = round(runif(1, 0, 10000)))
}
\arguments{
\item{msg}{A message object.}

\item{path}{Paths to files.}

\item{cid}{a content id.}
}
\value{
A message object.
}
\description{
Add attachments to a message object
}
\examples{
\dontrun{
msg <- envelope()
attachment(msg, "report.xlsx")
attachment(msg, c("visualisations.png", "report.pdf"))
}
}
