% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachments}
\alias{attachments}
\title{Add a multiple attachments to a message object}
\usage{
attachments(msg, paths)
}
\arguments{
\item{paths}{}
}
\description{
Add a multiple attachments to a message object
}
