% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body}
\alias{body}
\title{Add a body to a message object}
\usage{
body(msg, content = NULL, type = c("plain", "html"))
}
\arguments{
\item{msg}{A message object.}

\item{content}{A string of message content.}

\item{type}{Content type (plain text or HTML).}
}
\value{
A message object.
}
\description{
Add a body to a message object
}
\examples{
msg <- envelope()
body(msg, "Hello!")
body(msg, "<b>Hello!</b>", "html")
}
