% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{c.address}
\alias{c.address}
\title{Concatenate address objects}
\usage{
\method{c}{address}(...)
}
\arguments{
\item{...}{Address objects to be concatenated.}
}
\value{
An \code{address} object.
}
\description{
Concatenate address objects
}
\examples{
gerry <- as.address("Gerald <gerry@gmail.com>")
alice <- address("alice@yahoo.com")
jim <- address("jim@aol.com", "Jim")
c(gerry, alice)
c(gerry, c(alice, jim))
}
