% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-mail.R
\name{comments}
\alias{comments}
\title{Add or query comments of message.}
\usage{
comments(msg, comments = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{comments}{Comments for the message.}
}
\value{
A message object or the comments of the message object (if \code{comments} is \code{NULL}).
}
\description{
Add or query comments of message.
}
\examples{
# Create a message and set the comments.
msg <- envelope() \%>\% comments("This is a comment")

# Retrieve the comments for a message.
comments(msg)
}
\seealso{
\code{\link{subject}}
}
