% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare}
\alias{compare}
\title{Compare vectors}
\usage{
compare(lhs, rhs)
}
\arguments{
\item{lhs}{LHS of operation.}

\item{rhs}{RHS of operation.}
}
\value{

}
\description{
Returns \code{TRUE} wherever elements are the same (including \code{NA}),
and \code{FALSE} everywhere else.
}
