% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{display.address}
\alias{display.address}
\title{Extract display name}
\usage{
\method{display}{address}(address)
}
\arguments{
\item{address}{An \code{address} object.}
}
\value{
The display name or \code{NA}.
}
\description{
Extracts the display name from an email address.
}
