% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{display}
\alias{display}
\title{Extract display name}
\usage{
display(addr)
}
\arguments{
\item{addr}{An \code{address} object.}
}
\value{
The display name or \code{NA}.
}
\description{
Extracts the display name from an email address.
}
\examples{
gerry <- as.address("Gerald <gerry@gmail.com>")
display(gerry)
}
