% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{domain}
\alias{domain}
\title{Extract domain of email address}
\usage{
domain(addr)
}
\arguments{
\item{addr}{An \code{address} object.}
}
\value{
A character vector.
}
\description{
Extract domain of email address
}
\examples{
domain("alice@example.com")
}
