% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt message}
\usage{
encrypt(msg, sign = TRUE)
}
\arguments{
\item{msg}{A message object.}

\item{sign}{Whether to sign the message.}
}
\value{
A message object.
}
\description{
Encrypt message
}
