% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R, R/attachment.R, R/body.R,
%   R/envelope.R, R/message.R, R/subject.R
\name{to.envelope}
\alias{to.envelope}
\alias{from.envelope}
\alias{cc.envelope}
\alias{bcc.envelope}
\alias{attachment}
\alias{body.envelope}
\alias{envelope}
\alias{header.envelope}
\alias{message.envelope}
\alias{subject.envelope}
\title{Create a message object}
\usage{
\method{to}{envelope}(msg, ...)

\method{from}{envelope}(msg, from = NULL)

\method{cc}{envelope}(msg, ...)

\method{bcc}{envelope}(msg, ...)

attachment(msg, path)

\method{body}{envelope}(msg, content = NULL, type = c("plain", "html"))

envelope()

\method{header}{envelope}(msg)

\method{message}{envelope}(msg)

\method{subject}{envelope}(msg, topic = NULL)
}
\description{
Create a message object
}
\concept{envelope}
