% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{format.address}
\alias{format.address}
\title{Encode email addresses in a common format}
\usage{
\method{format}{address}(x, quote = TRUE, ...)
}
\arguments{
\item{x}{An \code{address} object.}

\item{quote}{Whether to quote display name (only relevant if display name is
given in "Last, First" format).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
Encode email addresses in a common format
}
