% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{format.mime}
\alias{format.mime}
\title{Format the header of a MIME object.}
\usage{
\method{format}{mime}(msg)
}
\arguments{
\item{msg}{A message object.}
}
\value{
A formatted header string.
}
\description{
The header must conform to the SMTP Protocol (https://tools.ietf.org/html/rfc821). Specifically, lines should be
terminated by CRLF (not just LF).
}
