% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{format.vctrs_address}
\alias{format.vctrs_address}
\title{Encode email addresses in a common format}
\usage{
\method{format}{vctrs_address}(x, ...)
}
\arguments{
\item{x}{A vector of \code{address} objects.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
Encode email addresses in a common format
}
