% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{html}
\alias{html}
\title{Add an HTML body to a message object}
\usage{
html(msg, content = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{content}{A string of message content.}
}
\value{
A message object.
}
\description{
Add an HTML body to a message object
}
\examples{
msg <- envelope()
html(msg, "<b>Hello!</b>")
}
