% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-mail.R
\name{id}
\alias{id}
\title{Set message ID.}
\usage{
id(msg, id)
}
\arguments{
\item{msg}{A message object.}

\item{id}{An ID for the message.}
}
\value{
A message object.
}
\description{
Set message ID.
}
\examples{
# Create a message and set the ID
msg <- envelope() \%>\% id("1234567890.123456@example.com")
}
