% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{length.address}
\alias{length.address}
\title{Length of address object}
\usage{
\method{length}{address}(x)
}
\arguments{
\item{x}{An \code{address} object.}
}
\value{
A character vector.
}
\description{
Length of address object
}
