% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{local}
\alias{local}
\title{Extract local part of email address}
\usage{
local(addr)
}
\arguments{
\item{addr}{An \code{address} object.}
}
\value{
A character vector.
}
\description{
Extract local part of email address
}
\examples{
local("alice@example.com")
}
