% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{md}
\alias{md}
\title{Render plain markdown into email}
\usage{
md(msg, input, interpolate = TRUE, .open = "{{", .close = "}}")
}
\arguments{
\item{msg}{A message object.}

\item{input}{The input markdown file to be rendered or a character vector of markdown text.}

\item{interpolate}{Whether to interpolate using \code{glue}.}

\item{.open}{The opening delimiter.}

\item{.close}{The closing delimiter.}
}
\value{
A message object.
}
\description{
Render plain markdown into email
}
\examples{
markdown <- "[This](https://www.google.com) is a link."
filename <- "message.md"

# Render from markdown in character vector.
msg <- envelope() \%>\% md(markdown)

# Create a file containing markdown.
cat(markdown, file = filename)

# Render from markdown in file.
msg <- envelope() \%>\% md(filename)

# Cleanup.
file.remove(filename)
}
