% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_id}
\alias{message_id}
\title{Create a message ID}
\usage{
message_id(domain = "mail.gmail.com")
}
\arguments{
\item{domain}{Originating domain.}
}
\value{
A message ID.
}
\description{
Create a message ID
}
\examples{
message_id()
message_id("example.com")
}
