% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{MIME}
\alias{MIME}
\title{Create a MIME object}
\usage{
MIME(
  content = NULL,
  disposition = NA,
  charset = NA,
  encoding = NA,
  boundary = emayili:::hexkey(),
  type = NA,
  children = list()
)
}
\arguments{
\item{content}{Content.}

\item{disposition}{Should the content be displayed inline or as an attachment?}

\item{charset}{How to interpret the characters in the content. Most often either UTF-8 or ISO-8859-1.}

\item{encoding}{How to encode binary data to ASCII.}

\item{boundary}{Boundary string.}

\item{type}{The MIME type of the content.}

\item{children}{}
}
\value{
A MIME object.
}
\description{
\preformatted{MIME
 ├── multipart/mixed
 ├── multipart/related
 ├── text/plain
 ├── text/html
 └── other
}
}
