% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{new_address}
\alias{new_address}
\title{Helper function for creating address objects}
\usage{
new_address(email = character(), display = character(), normalise = TRUE)
}
\arguments{
\item{email}{Email address.}

\item{display}{Display name.}

\item{normalise}{Whether to normalise address to RFC-5321 requirements.}
}
\value{
An \code{address} object, representing an email address.
}
\description{
Helper function for creating address objects
}
