% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{normalise.address}
\alias{normalise.address}
\title{Normalise email address}
\usage{
\method{normalise}{address}(address)
}
\arguments{
\item{address}{An \code{address} object.}
}
\value{

}
\description{
Makes an email address conform to RFC-5321.
}
