% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{normalise}
\alias{normalise}
\title{Normalise email address}
\usage{
normalise(address)
}
\arguments{
\item{address}{An \code{address} object.}
}
\value{
An RFC-5321 email address.
}
\description{
Makes an email address conform to RFC-5321.
}
\examples{
gerry <- as.address("     Gerald   Durrell   <   gerry@gmail.com  >   ")
normalise(gerry)
}
