% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{other}
\alias{other}
\title{Title}
\usage{
other(
  filename,
  name = NA,
  type = NA,
  cid = NA,
  disposition = "attachment",
  encoding = "base64",
  ...
)
}
\arguments{
\item{cid}{An optional Content-Id.}

\item{...}{}
}
\description{
Title
}
