% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties.R
\name{parties}
\alias{parties}
\title{Extract sender and recipient(s)}
\usage{
parties(msg)
}
\arguments{
\item{msg}{A message object.}
}
\value{
A tibble.
}
\description{
Extract sender and recipient(s)
}
\examples{
email <- envelope() \%>\%
  from("Gerald <gerald@gmail.com>") \%>\%
  to(c("bob@gmail.com", "alice@yahoo.com")) \%>\%
  cc("Craig     < craig@gmail.com>") \%>\%
  bcc("  Erin   <erin@yahoo.co.uk    >")

parties(email)
}
