% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{print.envelope}
\alias{print.envelope}
\title{Print a message object}
\usage{
\method{print}{envelope}(x, details = FALSE, ...)
}
\arguments{
\item{x}{A message object.}

\item{details}{Whether or not to display full message content.}

\item{...}{Any other arguments (for consistency of generic function).}
}
\description{
Print a message object
}
