% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{qp_decode}
\alias{qp_decode}
\title{Decode a quoted-printable string}
\usage{
qp_decode(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
A vector of decoded strings
}
\description{
Decode a quoted-printable string
}
