% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{raw.address}
\alias{raw.address}
\title{Extract raw email address}
\usage{
\method{raw}{address}(address)
}
\arguments{
\item{address}{An \code{address} object.}
}
\value{
A raw email address.
}
\description{
Strips the display name off an email address (if present).
}
