% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{raw}
\alias{raw}
\title{Extract raw email address}
\usage{
raw(addr)
}
\arguments{
\item{addr}{An \code{address} object.}
}
\value{
A raw email address.
}
\description{
Strips the display name off an email address (if present).
}
\examples{
gerry <- as.address("Gerald <gerry@gmail.com>")
raw(gerry)
}
