% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-receipt.R
\name{receipt}
\alias{receipt}
\alias{request_receipt_read}
\alias{request_receipt_delivery}
\title{Request read or delivery receipts}
\usage{
request_receipt_read(msg, addr = NULL)

request_receipt_delivery(msg, addr = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{addr}{Single address (optional). If \code{address} is not specified then
will use sender address.}
}
\value{
A message object.
}
\description{
Request the recipient to acknowledge that they have read the message. Inserts
MDN (Message Disposition Notification) header entries.
}
