% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{reply}
\alias{reply}
\title{Add Reply-To field to message}
\usage{
reply(msg, reply_to = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{reply_to}{Email address.}
}
\value{
A message object.
}
\description{
Add Reply-To field to message
}
\examples{
msg <- envelope()
reply(msg, "gerry@gmail.com")
}
\seealso{
\code{\link{to}}, \code{\link{cc}}, \code{\link{bcc}}, \code{\link{from}}, \code{\link{sender}} and \code{\link{subject}}
}
