% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-receipt.R
\name{request_receipt_read}
\alias{request_receipt_read}
\title{Request read receipt}
\usage{
request_receipt_read(msg, addr = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{addr}{Single address.}
}
\value{
A message object.
}
\description{
Request the recipient to acknowledge that they have read the message. Inserts
MDN (Message Disposition Notification) header entries.
}
