% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{rmd}
\alias{rmd}
\title{Render R markdown into email}
\usage{
rmd(msg, input, interpolate = TRUE, .open = "{{", .close = "}}")
}
\arguments{
\item{msg}{A message object.}

\item{input}{The input R markdown file to be rendered or a character vector of R markdown text.}

\item{interpolate}{Whether to interpolate using \code{glue}.}

\item{.open}{The opening delimiter.}

\item{.close}{The closing delimiter.}
}
\value{
A message object.
}
\description{
Render R markdown into email
}
\examples{
filename <- "gh-doc.Rmd"

# Create an Rmd document from template.
rmarkdown::draft(
  filename,
  template = "github_document",
  package = "rmarkdown",
  edit = FALSE
)

# Render from Rmd file.
msg <- envelope() \%>\% rmd(filename)

# Cleanup.
file.remove(filename)
}
