% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{sender}
\alias{sender}
\title{Add Sender (on behalf of) field to message}
\usage{
sender(msg, sender = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{sender}{Email address.}
}
\value{
A message object.
}
\description{
Add Sender (on behalf of) field to message
}
\examples{
msg <- envelope()
sender(msg, "on_behalf_of@gmail.com")
}
\seealso{
\code{\link{to}}, \code{\link{cc}}, \code{\link{bcc}}, \code{\link{from}}, \code{\link{reply}} and \code{\link{subject}}
}
