% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\title{Add or query subject of message}
\usage{
subject(msg, subject = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{subject}{A subject for the message.}
}
\value{
A message object or the subject of the message object (if \code{subject} is \code{NULL}).
}
\description{
Add or query subject of message
}
\examples{
msg <- envelope()
subject(msg, "Updated report")
}
