% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{text}
\alias{text}
\title{Add a text body to a message object}
\usage{
text(msg, content = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{content}{A string of message content.}
}
\value{
A message object.
}
\description{
Add a text body to a message object
}
\examples{
msg <- envelope()
text(msg, "Hello!")
}
