% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{text_html}
\alias{text_html}
\title{Create \code{text/html} MIME object}
\usage{
text_html(
  content,
  disposition = "inline",
  charset = "utf-8",
  encoding = NA,
  ...
)
}
\arguments{
\item{content}{Content.}

\item{disposition}{Should the content be displayed inline or as an attachment?}

\item{charset}{How to interpret the characters in the content. Most often either UTF-8 or ISO-8859-1.}

\item{encoding}{How to encode binary data to ASCII.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{}text_html object derived from \code{MIME}.
}
\description{
Create \code{text/html} MIME object
}
