% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{validate}
\alias{validate}
\title{Validate email address}
\usage{
validate(addr, deliverability = TRUE)
}
\arguments{
\item{addr}{An email address.}

\item{deliverability}{Whether to check for deliverability (valid domain).}
}
\value{
A \code{logical} indicating whether or not the address is valid.
}
\description{
Validate email address
}
\examples{
# A valid address.
validate("cran-sysadmin@r-project.org")
# An invalid address.
validate("help@this-domain-does-not-exist.com")
}
