% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{vec_ptype_abbr.vctrs_address}
\alias{vec_ptype_abbr.vctrs_address}
\title{Display abbreviated type of vector}
\usage{
vec_ptype_abbr.vctrs_address(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Display abbreviated type of vector
}
