% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{address}
\alias{address}
\title{Email Address}
\usage{
address(email, display = NA, normalise = TRUE)
}
\arguments{
\item{email}{Email address.}

\item{display}{Display name.}
}
\value{
An \code{address} object, representing an email address.
}
\description{
Implemented as an \href{https://cran.r-project.org/web/packages/vctrs/vignettes/s3-vector.html}{S3 vector class}.
}
\examples{
address("gerry@gmail.com")
address("gerry@gmail.com", "Gerald")
address(
  c("gerry@gmail.com", "alice@yahoo.com", "jim@aol.com", NA),
  c("Gerald", "Alice", NA, "Bob")
)
}
