% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{as.address}
\alias{as.address}
\title{Create an address object}
\usage{
as.address(addr)
}
\arguments{
\item{addr}{An email address.}

\item{split}{Pattern for splitting multiple addresses.}
}
\value{
A list of \code{address} objects.
}
\description{
Create an address object
}
\examples{
as.address("gerry@gmail.com")
as.address("Gerald <gerry@gmail.com>")
as.address(c("Gerald <gerry@gmail.com>", "alice@yahoo.com", "jim@aol.com"))
as.address("Gerald <gerry@gmail.com>, alice@yahoo.com, jim@aol.com")
as.address("Durrell, Gerald <gerry@gmail.com>")
}
