% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{as.character.address}
\alias{as.character.address}
\title{Convert address object to character}
\usage{
\method{as.character}{address}(x, ...)
}
\arguments{
\item{x}{An \code{address} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
If display name is specifed as "Last, First" then the display name will be
quoted.
}
