% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{as.character.header}
\alias{as.character.header}
\title{Create formatted header.}
\usage{
\method{as.character}{header}(x, width = 30, ...)
}
\arguments{
\item{x}{A header object.}

\item{width}{The width of the head name field.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A formatted header field.
}
\description{
Accepts a header object and formats it as a header field.
}
