% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add an attachment to a message object}
\usage{
attachment(msg, path)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Path to a file.}
}
\value{
A message object.
}
\description{
Add an attachment to a message object
}
\examples{
msg \%>\% attachment("report.xlsx") \%>\% attachment("visualisations.png")
}
