% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add a single attachment to a message object}
\usage{
attachment(msg, path, cid = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Paths to files.}

\item{cid}{Associate a Content ID.}
}
\value{
A message object.
}
\description{
Add a single attachment to a message object
}
\examples{
\dontrun{
msg <- envelope()
attachment(msg, "report.xlsx")
attachment(msg, c("visualisations.png", "report.pdf"))
attachment(msg, c("visualisations.png", "image.jpg"), cid = TRUE)
}
}
