% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add attachments to a message object}
\usage{
attachment(msg, path, type = NA, cid = NA, disposition = NA)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Path to file.}

\item{type}{MIME type or \cite{NA}, which will result in a guess based on file extension.}

\item{cid}{Content-ID or \code{NA}.}

\item{disposition}{Should the content be displayed inline or as an attachment?}
}
\value{
A message object.
}
\description{
Add attachments to a message object
}
\examples{
library(magrittr)

download.file("https://bit.ly/2P4LUO8", "cats.jpg")

png("scatter.png")
plot(1:10)
dev.off()

write.csv(mtcars, "mtcars.csv")

msg <- envelope() \%>\%
  attachment("mtcars.csv") \%>\%
  attachment("cats.jpg", type = "image/jpeg") \%>\%
  attachment("scatter.png", cid = "scatter")

file.remove("mtcars.csv", "scatter.png", "cats.jpg")
}
