% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add attachments to a message object}
\usage{
attachment(msg, path, cid = NA, disposition = NA, mime_type)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Path to file.}

\item{cid}{Content-ID or \code{NA}.}

\item{mime_type}{fixed mime type without guessing. Use \link[mime]{mimemap} to set. (optional)}
}
\value{
A message object.
}
\description{
Add attachments to a message object
}
\examples{
\dontrun{
msg <- envelope()
attachment(msg, "report.xlsx")
attachment(msg, "cat.png")
attachment(msg, "visualisations.png", "visuals")
}
}
