% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add attachments to a message object}
\usage{
attachment(msg, path, cid = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Paths to files.}

\item{cid}{A vector of the same length of path containing either NA or string values specifying a Content ID for each attachment.}
}
\value{
A message object.
}
\description{
Add attachments to a message object
}
\examples{
\dontrun{
msg <- envelope()
attachment(msg, "report.xlsx")
attachment(msg, c("visualisations.png", "report.pdf"), c("visualizations_01", NA))
}
}
