% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment.R
\name{attachment}
\alias{attachment}
\title{Add attachments to a message object}
\usage{
attachment(msg, path, type = NA, cid = NA, disposition = NA)
}
\arguments{
\item{msg}{A message object.}

\item{path}{Path to file.}

\item{type}{MIME type or \cite{NA}, which will result in a guess based on file extension.}

\item{cid}{Content-ID or \code{NA}.}
}
\value{
A message object.
}
\description{
Add attachments to a message object
}
\examples{
\dontrun{
msg <- envelope()
attachment(msg, "report.xlsx")
attachment(msg, "cat.png", type = "image/png")
attachment(msg, "visualisations.png", "visuals")
}
}
