% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{bcc}
\alias{bcc}
\title{Add Bcc field to message}
\usage{
bcc(msg, ...)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email addresses.}
}
\value{
A message object.
}
\description{
Add Bcc field to message
}
\examples{
msg <- envelope()
bcc(msg, "bob@gmail.com", "alice@yahoo.com")
bcc(msg, c("bob@gmail.com", "alice@yahoo.com"))
}
\seealso{
\code{\link{subject}}, \code{\link{from}}, \code{\link{to}}, \code{\link{cc}} and \code{\link{reply}}
}
