% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{bcc}
\alias{bcc}
\title{Add Bcc field to message object header}
\usage{
bcc(msg, ...)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email addresses.}
}
\value{
A message object.
}
\description{
Add Bcc field to message object header
}
\examples{
envelope() \%>\% bcc("bob@gmail.com", "alice@yahoo.com")
envelope() \%>\% bcc(c("bob@gmail.com", "alice@yahoo.com"))
}
