% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties.R
\name{cc}
\alias{cc}
\title{Add Cc field to message}
\usage{
cc(msg, ...)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email addresses.}
}
\value{
A message object.
}
\description{
Add Cc field to message
}
\examples{
msg <- envelope()
cc(msg, "bob@gmail.com", "alice@yahoo.com")
cc(msg, c("bob@gmail.com", "alice@yahoo.com"))
}
\seealso{
\code{\link{to}}, \code{\link{bcc}}, \code{\link{from}}, \code{\link{sender}}, \code{\link{reply}} and \code{\link{subject}}
}
