% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{cleave}
\alias{cleave}
\title{Split into individual addresses}
\usage{
cleave(addr)
}
\arguments{
\item{addr}{}
}
\value{

}
\description{
Split into individual addresses
}
\examples{
to <- as.address(c("Gerald <gerry@gmail.com>", "alice@yahoo.com", "jim@aol.com"))
cleave(to)
}
