% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{compliant}
\alias{compliant}
\title{Tests whether an email address is syntactically correct}
\usage{
compliant(addr)
}
\arguments{
\item{addr}{An email address.}
}
\value{
A Boolean.
}
\description{
Tests whether an email address is syntactically correct
}
\examples{
compliant("alice@example.com")
compliant("alice?example.com")
}
