% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt message}
\usage{
encrypt(msg, sign = TRUE)
}
\arguments{
\item{msg}{A message object.}

\item{sign}{Whether to sign the message.}
}
\value{
A message object.
}
\description{
Encrypt message
}
\examples{
#'
\dontrun{
# A WWII D-Day message.
#
# See https://twitter.com/bletchleypark/status/1136490396626800640.
#
msg <- envelope(
  to = "schunk@u-boat.com",
  subject = "Top Secret Message",
  text = "Immediate readiness. There are indications that the invasion has begun."
)
msg \%>\% encrypt()
}
}
