% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Create a message.}
\usage{
envelope()
}
\value{
A message object.
}
\description{
Create a message.
}
\examples{
# Create an (empty) message object.
msg <- envelope()
}
\seealso{
\code{\link{subject}}, \code{\link{from}}, \code{\link{to}}, \code{\link{cc}}, \code{\link{bcc}} and \code{\link{reply}}
}
