% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{from}
\alias{from}
\title{Add From field to message object header}
\usage{
from(msg, from = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email address.}
}
\value{
A message object.
}
\description{
Add From field to message object header
}
\examples{
msg <- envelope()
from(msg, "craig@gmail.com")
}
