% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message}
\alias{message}
\title{Create formatted message.}
\usage{
message(msg)
}
\arguments{
\item{msg}{A message object.}
}
\value{
A formatted message object.
}
\description{
Create formatted message.
}
