% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{mime}
\alias{mime}
\alias{format.mime}
\title{Create a MIME (Multipurpose Internet Mail Extensions) object}
\usage{
mime(content_type, encoding, format, charset, ...)

\method{format}{mime}(msg)
}
\description{
Create a MIME (Multipurpose Internet Mail Extensions) object
}
