% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime.R
\name{mime}
\alias{mime}
\title{Create a MIME (Multipurpose Internet Mail Extensions) object}
\usage{
mime(content_type, encoding, format, charset, ...)
}
\value{
A MIME object.
}
\description{
Create a MIME (Multipurpose Internet Mail Extensions) object
}
