% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{normalise}
\alias{normalise}
\title{Normalise email address}
\usage{
normalise(email)
}
\arguments{
\item{email}{An email address.}
}
\value{
An email address.
}
\description{
Ensure that email address is in a standard format.
}
\details{
Performs the following transformations:
\itemize{
\item lowercase the domain part
\item replace some Unicode characters with compatible equivalents. See
\href{https://en.wikipedia.org/wiki/Unicode_equivalence}{Unicode equivalence}.
}
}
\examples{
normalise("bob@GMAIL.COM")
\dontshow{
# This breaks creation of the PDF manual.
normalise("bob@ｇｍａｉｌ.com")
}
}
