% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{print.address}
\alias{print.address}
\title{Print an address object}
\usage{
\method{print}{address}(address)
}
\arguments{
\item{address}{An \code{address} object.}
}
\description{
Print an address object
}
\examples{
gerry <- as.address("gerry@gmail.com")
print(gerry)
}
