% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{print.envelope}
\alias{print.envelope}
\title{Print a message object}
\usage{
\method{print}{envelope}(x, details = NA, ...)
}
\arguments{
\item{x}{A message object.}

\item{details}{Whether or not to display full message content.}

\item{...}{Any other arguments (for consistency of generic function).}
}
\description{
The message body will be printed if \code{details} is \code{TRUE} or if the \code{envelope_details}
option is \code{TRUE}.
}
\examples{
msg <- envelope() \%>\% text("Hello, World!")

print(msg)
print(msg, details = TRUE)

options(envelope_details = TRUE)
print(msg)
}
