% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{qp}
\alias{qp}
\alias{qp_encode}
\alias{qp_decode}
\title{Quoted-Printable encoding}
\usage{
qp_encode(x, crlf = CRLF)

qp_decode(x)
}
\arguments{
\item{x}{A string for encoding or decoding.}

\item{crlf}{End-of-line characters.}
}
\value{
An encoded string for \code{qp_encode()} or a decoded string for \code{qp_decode()}.
}
\description{
Encode to and decode from Quoted-Printable encoding.
}
\examples{
qp_encode("Mieux vaut être seul que mal accompagné.")
qp_decode("Mieux vaut =C3=AAtre seul que mal accompagn=C3=A9.")
}
