% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render markdown into email}
\usage{
render(msg, input)
}
\arguments{
\item{msg}{A message object.}

\item{input}{The input markdown file to be rendered or a character vector of markdown text.}
}
\value{
A message object.
}
\description{
Render markdown into email
}
\examples{
md <- "[This](https://www.google.com) is a link."
filename <- "message.md"

# Render from markdown in character vector.
msg <- envelope() \%>\% render(md)

# Create a file containing markdown.
cat(md, file = filename)

# Render from markdown in file.
msg <- envelope() \%>\% render(filename)

# Cleanup.
file.remove(filename)
}
