% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render R Markdown document into email}
\usage{
render(msg, rmd_filepath, quiet = TRUE)
}
\arguments{
\item{msg}{A message object.}

\item{rmd_filepath}{Path to a Rmd file.}

\item{quiet}{Suppress output.}
}
\value{

}
\description{
Render R Markdown document into email
}
