% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{reply}
\alias{reply}
\title{Add Reply-To field to message object header}
\usage{
reply(msg, reply_to = NULL)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email addresses.}
}
\value{
A message object.
}
\description{
Add Reply-To field to message object header
}
\examples{
msg <- envelope()
reply(msg, "gerry@gmail.com")
}
