% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Create a SMTP server object}
\usage{
server(host, port = 25, username = NULL, password = NULL)
}
\arguments{
\item{host}{DNS name or IP address of the SMTP server.}

\item{port}{Port that the SMTP server is listening on.}

\item{username}{Username for SMTP server.}

\item{password}{Password for SMTP server.}
}
\value{
A function which is used to send messages to the server.
}
\description{
Create a SMTP server object
}
\examples{
library(magrittr)

smtp <- server(host = "smtp.gmail.com",
               port = 465,
               username = "bob@gmail.com",
               password = "bd40ef6d4a9413de9c1318a65cbae5d7")
msg <- envelope() \%>\%
  from("bob@gmail.com") \%>\%
  to("alice@yahoo.com")
\dontrun{
smtp(msg, verbose = TRUE)
}
}
