% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{text}
\alias{text}
\title{Add a text body to a message object}
\usage{
text(
  msg,
  content,
  disposition = "inline",
  charset = "utf-8",
  encoding = "7bit"
)
}
\arguments{
\item{msg}{A message object.}

\item{content}{A string of message content.}

\item{disposition}{How content is presented (Content-Disposition).}

\item{charset}{How content is encoded.}

\item{encoding}{How content is transformed to ASCII (Content-Transfer-Encoding).}
}
\value{
A message object.
}
\description{
Add a text body to a message object
}
\examples{
msg <- envelope()
text(msg, "Hello!")
}
