% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{to}
\alias{to}
\title{Add To field to message}
\usage{
to(msg, ...)
}
\arguments{
\item{msg}{A message object.}

\item{...}{Email addresses.}
}
\value{
A message object.
}
\description{
Add To field to message
}
\examples{
msg <- envelope()
to(msg, "bob@gmail.com", "alice@yahoo.com")
to(msg, c("bob@gmail.com", "alice@yahoo.com"))
}
\seealso{
\code{\link{subject}}, \code{\link{from}}, \code{\link{cc}}, \code{\link{bcc}}, \code{\link{reply}}, and \code{\link{reply}}
}
