% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{vec_ptype_abbr.vctrs_address}
\alias{vec_ptype_abbr.vctrs_address}
\title{Display abbreviated type of vector}
\usage{
vec_ptype_abbr.vctrs_address(x, ...)
}
\description{
Display abbreviated type of vector
}
